refactor: move prettyBytes into vendor/ with upstream LICENSE
